Update to HF production, using MDC1.4
condor submission script, run_all.pl, has been updated to generate 2000 events per job. The perl command should then be `perl run_all.pl 25000 Charm` to generate 50M c-cbar events (switch Charm for Bottom to generate the equivalent b-bbar events). This is 25k jobs x 2e3 events/job = 50M events

run_hfprod.sh script has been updated to source the MDC 1.4 build with KFParticle

Fun4All_G4_HF_pp_signal.C has a few updates:

(1) IMPORTANT! We will use the QA output from our production. This originally would be written to MDC1/submit/HF_pp200_signal/rundir/ and use about 1.2TB of storage. To avoid this, I have now sent the QA files to the same location as the production files BUT they will be in a sub-folder called "QA". If this folder does not exit, then Fun4All_G4_HF_pp_signal.C will create it. This might not be behavior we want but it was a way for me to separate production and QA

(2) Jet QA has been disabled as it still causes a crash with truth matching. There is a random seed in the macro which, if uncommented, will reproduce the problem.

(3) The track-vertex DCA chi2 has been set to 0 for our containers. We now reconstruct SV as well as PV in the production and KFParticle cannot tell the difference between the two, containers will be empty if this paameter is non-zero for now. This will be updated at a later date but does not impact the production. I have two ideas on how to recover this functionality.

G4_KFParticle.C has been removed as it is now picked up by sourcing the MDC build

Note, dcache files have not been tested